HP Bar and simple Explosin indicator.
Boss HP bar changes color based on Phase.
Bomber's Bar rotates around as bomber rotates, but I thought it was funny enough to leave it as is.. and fixing it is bit annoying....

Also tried AoE(Bomber Special) but I Couldn't get it to work right.
from few trial run, it seems shield failing might be enough of incentive to move around.